Chapter 9 Updates
 - Updated class names to match later chapter standards.
 - Formatted files.
 - Updated settings to make sure 4 spaces are preferred when indenting.
 - Added new reason to miss points on an assignemnt.
