+++ SECURITY.md
@@ -8,13 +8,13 @@
 
 ## Reporting a Vulnerability
 
-To report a vulnerability with the code, please email Patrick.McDougle@gmail with the information needed.  Please provide the following:
+To report a vulnerability with the code, please email Patrick.McDougle@gmail with the information needed. Please provide the following:
 
-  * File Name
-  * Line(s) where the vulnerability is located
-  * Why it is a vulnerability.
-  * Any supporting online documentation on the vulnerability.
+-   File Name
+-   Line(s) where the vulnerability is located
+-   Why it is a vulnerability.
+-   Any supporting online documentation on the vulnerability.
 
 ## Updates on reported vulnerability
 
-Feedback on the vulnerability (course of action, time line, ect...) will be sent to the reporting parties as soon as possible.  This maybe be in a week or a month depending on the issue and availability to look into this.
+Feedback on the vulnerability (course of action, time line, ect...) will be sent to the reporting parties as soon as possible. This maybe be in a week or a month depending on the issue and availability to look into this.
