--- src/edu/csc1061/ch09/Ch09Circle.java
@@ -1,40 +1,0 @@
-/**
- * Code for Class.
- * <p>
- * CSC 1061 - Computer Science II - Java
- *
- * @author  Patrick McDougle
- * @version %I%, %G%
- * @since   1.0
- */
-package edu.csc1061.ch09;
-
-// Define the circle class with two constructors
-class Ch09Circle {
-  double radius;
-
-  /** Construct a circle with radius 1 */
-  Ch09Circle() {
-    radius = 1;
-  }
-
-  /** Construct a circle with a specified radius */
-  Ch09Circle(double newRadius) {
-    radius = newRadius;
-  }
-
-  /** Return the area of this circle */
-  double getArea() {
-    return radius * radius * Math.PI;
-  }
-
-  /** Return the perimeter of this circle */
-  double getPerimeter() {
-    return 2 * radius * Math.PI;
-  }
-
-  /** Set a new radius for this circle */
-  void setRadius(double newRadius) {
-    radius = newRadius;
-  }
-}
