+++ src/edu/csc1061/ch09/Ch09Program03.java
@@ -1,0 +1,19 @@
+/**
+ * Code for Class.
+ * <p>
+ * CSC 1061 - Computer Science II - Java
+ *
+ * @author Patrick McDougle
+ * @version %I%, %G%
+ * @since 1.0
+ */
+package edu.csc1061.ch09;
+
+public class Ch09Program03 {
+
+    public static void main(String[] args) {
+        Student2 student = new Student2(111223333, 1970, 5, 3);
+        BirthDate date = student.getBirthDate();
+        date.setYear(2010); // Now the student birth year is changed!
+    }
+}
