+++ src/edu/csc1061/ch09/Circle2.java
@@ -9,18 +9,29 @@
  */
 package edu.csc1061.ch09;
 
-public class Ch09Circle1 {
+public class Circle2 {
     // the radius of this circle
-    public double radius;
+    public double radius; // instance variable
+
+    // static variables are shared by all the instances of the class.
+    // the number of objects created
+    public static int numberOfObjects = 0;
 
     // Construct a circle with radius 1
-    public Ch09Circle1() {
+    public Circle2() {
         radius = 1;
+        numberOfObjects++;
     }
 
     // Construct a circle with a specified radius
-    public Ch09Circle1(double newRadius) {
+    public Circle2(double newRadius) {
         radius = newRadius;
+        numberOfObjects++;
+    }
+
+    // Return numberOfObjects
+    static int getNumberOfObjects() {
+        return numberOfObjects;
     }
 
     // Return the radius of the circle
@@ -31,7 +42,7 @@ public class Ch09Circle1 {
     // Return the area of this circle
     // PI * radius^2
     public double getArea() {
-        return radius * radius * Math.PI;
+        return radius * numberOfObjects * Math.PI;
     }
 
     // Return the perimeter (circumference) of this circle
