+++ src/edu/csc1061/ch09/Student1.java
@@ -3,15 +3,15 @@
  * <p>
  * CSC 1061 - Computer Science II - Java
  *
- * @author  Patrick McDougle
+ * @author Patrick McDougle
  * @version %I%, %G%
- * @since   1.0
+ * @since 1.0
  */
 package edu.csc1061.ch09;
 
 public class Student1 {
-  String name; // default value null
-  int age; // default value 0
-  boolean isScienceMajor; // default value false
-  char gender; // default value '\u0000'
+    String name; // default value null
+    int age; // default value 0
+    boolean isScienceMajor; // default value false
+    char gender; // default value '\u0000'
 }
